requirements.txt
yt-dlp>=2024.1.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
ollama>=0.1.0
whisperx>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
